requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0.1
djangorestframework==3.16.1
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-dialogflow==2.45.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==24.1.1
idna==3.11
packaging==26.0
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
requests==2.32.5
rsa==4.9.1
schedule==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
